클라이언트와 통신 테스트 12/15

+++ Server/Models/processModel.cs
@@ -4,6 +4,6 @@
     {
         public string processCmd { get; set; }
         public string processName { get; set; }
-        public double processValue { get; set; }
+        public float processValue { get; set; }
     }
 }
